requirements.txt
openai
streamlit
faiss-cpu
numpy
tiktoken
google.generativeai